Day 7 - JS: Part 1 passing.
I realized that each wire reference was either going to be a wire name (i.e. a string of lowercase letters) or a number.

I can't perform a bitwise operation until all involved operands are numbers, and I similarly can't assign a wire a value unless it's a number.

This approach iterates through the list of instructions, splicing out any whose numerical result we are able to compute (i.e. whose wire names can all be resolved to numerical values) and skipping the rest, returning to the start of the list until all instructions have been processed.

I'm certain this code can be cleaned up substantially, but I'll do that as a separate commit.